rename: application service 폴더 생성 및 정리 #5
